+++ dbdaora/hash/query.py
@@ -4,15 +4,17 @@ from typing import Any, ClassVar, List, Optional, Sequence, Tuple, Type, Union
 from dbdaora.keys import FallbackKey
 from dbdaora.query import BaseQuery, Query, QueryMany
 
+from .repositories import HashData, HashEntity, HashRepository
+
 
 @dataclasses.dataclass(init=False)
-class HashQuery(Query[Any, 'HashData', FallbackKey]):
-    repository: 'HashRepository[FallbackKey]'
+class HashQuery(Query[HashEntity, HashData, FallbackKey]):
+    repository: HashRepository[HashEntity, FallbackKey]
     fields: Optional[Sequence[str]] = None
 
     def __init__(
         self,
-        repository: 'HashRepository[FallbackKey]',
+        repository: HashRepository[HashEntity, FallbackKey],
         *args: Any,
         memory: bool = True,
         key_parts: Optional[List[Any]] = None,
@@ -26,15 +28,17 @@ class HashQuery(Query[Any, 'HashData', FallbackKey]):
 
 
 @dataclasses.dataclass(init=False)
-class HashQueryMany(QueryMany[Any, 'HashData', FallbackKey]):
-    query_cls: ClassVar[Type[HashQuery[FallbackKey]]] = HashQuery[FallbackKey]
-    queries: Sequence[HashQuery[FallbackKey]]  # type: ignore
-    repository: 'HashRepository[FallbackKey]'
+class HashQueryMany(QueryMany[HashEntity, HashData, FallbackKey]):
+    query_cls: ClassVar[Type[HashQuery[HashEntity, FallbackKey]]] = HashQuery[
+        HashEntity, FallbackKey
+    ]
+    queries: Sequence[HashQuery[HashEntity, FallbackKey]]  # type: ignore
+    repository: HashRepository[HashEntity, FallbackKey]
     fields: Optional[Sequence[str]] = None
 
     def __init__(
         self,
-        repository: 'HashRepository[FallbackKey]',
+        repository: HashRepository[HashEntity, FallbackKey],
         *args: Any,
         many: List[Union[Any, Tuple[Any, ...]]],
         memory: bool = True,
@@ -56,11 +60,10 @@ class HashQueryMany(QueryMany[Any, 'HashData', FallbackKey]):
             query.fields = fields
 
 
-def make(*args: Any, **kwargs: Any) -> BaseQuery[Any, 'HashData', FallbackKey]:
+def make(
+    *args: Any, **kwargs: Any
+) -> BaseQuery[HashEntity, HashData, FallbackKey]:
     if kwargs.get('many') or kwargs.get('many_key_parts'):
         return HashQueryMany(*args, **kwargs)
 
     return HashQuery(*args, **kwargs)
-
-
-from .repositories import HashRepository, HashData  # noqa isort:skip