+++ dbdaora/hash/repositories/mongodb.py
@@ -1,7 +1,7 @@
 from dbdaora.data_sources.fallback.mongodb import Key
 
-from . import HashRepository
+from . import HashEntity, HashRepository
 
 
-class MongodbHashRepository(HashRepository[Key]):
+class MongodbHashRepository(HashRepository[HashEntity, Key]):
     __skip_cls_validation__ = ('MongodbHashRepository',)