+++ docs/src/index/simple_hash.py
@@ -19,8 +19,7 @@ def make_person(name: str, age: int) -> Person:
     return Person(name.replace(' ', '_').lower(), name, age)
 
 
-class PersonRepository(HashRepository[str]):
-    entity_cls = Person
+class PersonRepository(HashRepository[Person, str]):
     key_attrs = ('id',)